requirements.txt
streamlit
pandas
streamlit-cookies-controller
mysql-connector-python